fix: minor code refactor
