--- packages/client/src/modules/index.ts
@@ -1,27 +1,0 @@
-import { About } from "./About/About";
-import { ChatDetails } from "./ChatDetails/ChatDetails";
-import { ChatDetailsContent } from "./ChatDetails/ChatDetailsContent";
-import { History } from "./History/History";
-import { HistoryContent } from "./History/HistoryContent";
-import { MessageAssistant } from "./Messages/MessageAssistant";
-import { MessagesContainer } from "./Messages/MessagesContainer";
-import { MessagesContainerHeader } from "./Messages/MessagesContainerHeader";
-import { PromptInput } from "./Messages/PromptInput";
-import { Profile } from "./Profile/Profile";
-import { ProfileContent } from "./Profile/ProfileContent";
-import { Toolbox } from "./Toolbox/Toolbox";
-
-export {
-	About,
-	ChatDetails,
-	ChatDetailsContent,
-	History,
-	HistoryContent,
-	MessageAssistant,
-	MessagesContainer,
-	MessagesContainerHeader,
-	Profile,
-	ProfileContent,
-	PromptInput,
-	Toolbox,
-};
